Add files via upload
This MATLAB file loads Raman spectra stored in two text files, one for Parallel and one for Cross configuration, subtracts baseline using airPLS method, and plots the spectra.  
It also fits the Raman peaks of each spectrum to a Guassian or Lorentizan function and extracts the fitting data for each peak (center, amplitude, and FWHM ). Moreover, it fits the polar plots of a Raman mode. 
Note: you will need the airPLS and distinguishable_colors files in the same working directory for this code to work. These files are also attached.